requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Pillow>=10.1.0
pytesseract==0.3.10
pdfplumber==0.10.0
spacy==3.7.0
